Scr-server now sends an image to the client, tagged "img" and last in order. The image is huge, ~300000 pixels, each an unsigned char casted to int and appended to the string. This slows down the visualization to less then a pixel per second. Next commit will include a switch to activate the image generation from command line - BEWARE USING THIS ONE

+++ src/drivers/scr_server/SimpleParser.cpp
@@ -134,7 +134,7 @@ SimpleParser::stringify(string tag, unsigned char *value, int size)
 	STR << "(" << tag;
 	for (int i = 0; i < size; ++i) 
 	{
-		STR << " " << value[i];
+		STR << " " << (int)value[i];
 	}
 	STR << ")";
 	return STR.str();
